Feature: Persist Game Board State
Feat: Add empty game board

FSA2021-74

Test: Add test for empty game board

FSA2021-74

Feat: Add markBoard() method to Game object

FSA2021-74

Test: Move game tests to separate file and add test for markBoard() method

FSA2021-74

Feat: Update handleClickCell() function to include the markBoard() method

This also involved adding the index as an argument. This caused some previously written tests that involved
handleClickCell() to fail, so we updated each with an index to correct them.

FSA2021-74

Style: Fix linter error for extra empty line

FSA2021-74

Style: Add missing semicolons

FSA2021-74

Style: Break up long line

FSA2021-74

Style: Remove unused variable

FSA2021-74

refactor: remove handleClick
- We are removing handle click and placing it back in index.js
- switching to default export for linting

refactor: import Game and HCC
- Importing our Game.js file and the code for handleClickCell
- we are going to instantiate Game.js after importing it
- we use the Game instantiation in our paremeter

refactor: structured file

refactor: change import

refactor: moving handleClick
sorry for moving this again.... but i realized leaving handeClick in index.js will run code in our test environment that we do not  need. I am moving handleClickCell to its own file
- imported handleClickCell.js
- passed game in as a direct argument to handleClickCell

refactor: new handleClickCell js
- moved handleClick cell into its own file

refactor: linting

refactor: changed imports for tes

test: add ClickHandler test file
- moved all tests that were in index.spec.js file that had anything to do with handleClick
- removed any imports not needed

test: removed unused test
-removed unused tests from index.spec.js

Test: Prepare test suite

I prepared an empty test suite for three tests related to confirming that the changes
we make in the UI are also reflected in our game object.

FSA2021-74

Test: Add test for empty grid cells (UI) and empty game board (object)

FSA2021-74

Test: Set up addHandleClickListener in test block

FSA2021-74

Test: Add test for confirming marking X marks X in the UI and the game object

FSA2021-74

Test: Add test to confirm marking O changes both UI and game object

FSA2021-74

Style: Change "test" to "it" to conform with other test blocks

FSA2021-74

test: moved addedUI test
- moved added UI test to handleClickCell

style: arrow function styling
