+++ src/js/index.js
@@ -1,17 +1,16 @@
 /* eslint-disable import/extensions */
-import { handleClickCell } from './Game/game.js';
+import Game from './Game/game.js';
+import handleClickCell from './HandleClick/handleClickCell.js';
 import { updateVersionNumber } from './version.js';
-
-console.debug('Welcome to Tic Tac Toe!');
-
 // eslint-disable-next-line no-undef
 const subheader = document.querySelector('.subheader');
-// eslint-disable-next-line quotes
-subheader.innerText = (`It's X's turn`);
+subheader.innerText = ('It\'s X\'s turn');
 // eslint-disable-next-line no-undef
 const allGameCells = document.querySelectorAll('.grid-cell');
-allGameCells.forEach((cell) => cell.addEventListener('click', () => { handleClickCell(cell, subheader); }));
-
 const appVersion = document.querySelector('.app-version_version');
+const game = new Game();
 
+console.debug('Welcome to Tic Tac Toe!');
 updateVersionNumber(appVersion);
+
+allGameCells.forEach((cell, index) => cell.addEventListener('click', () => handleClickCell(cell, subheader, index, game)));
